requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
numpy>=1.24,<2.0
pandas>=2.0
pytesseract==0.3.10
opencv-python-headless==4.8.0.76
onnxruntime==1.22.0
pydantic>=2.0
Pillow==10.2.0